pyproject.toml
[project]
name = "ahe-whisper-v90"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "dacite>=1.9.2",
    "huggingface-hub>=1.1.6",
    "librosa>=0.10.0,<0.11.0",
    "mlx-whisper>=0.4.3",
    "nicegui>=1.4.0,<1.5.0",
    "numpy>=1.26.4,<2.0.0",
    "onnxruntime>=1.17.0,<1.19.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.3.0,<1.6.0",
    "scipy>=1.11.0,<1.14.0",
    "sentencepiece==0.2.1",
    "sherpa-onnx>=1.12.18",
    "soundfile>=0.12.0,<0.13.0",
    "sudachidict-core>=20251022",
    "sudachipy>=0.6.10",
]
